Bug fix - we were reporting more metrics than specified (even when they were not captured)
