+++ src/xunit.performance.api/MetricCollectionFactory/StopwatchPerformanceMetricFactory.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Xunit.Performance.Sdk;
 using System.Collections.Generic;
+using static Microsoft.Xunit.Performance.BenchmarkMetricDiscoverer;
 
 namespace Microsoft.Xunit.Performance.Api
 {
@@ -7,7 +8,7 @@ namespace Microsoft.Xunit.Performance.Api
     {
         public IEnumerable<PerformanceMetric> GetMetrics(string assemblyFileName)
         {
-            return new List<PerformanceMetric>();
+            return new List<PerformanceMetric> { new BenchmarkDurationMetric() };
         }
     }
 }
