+++ src/xunit.performance.metrics/BenchmarkDurationMetric.cs
@@ -1,0 +1,23 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.Xunit.Performance.Sdk;
+
+namespace Microsoft.Xunit.Performance
+{
+    internal partial class BenchmarkMetricDiscoverer : IPerformanceMetricDiscoverer
+    {
+        internal class BenchmarkDurationMetric : PerformanceMetric
+        {
+            public BenchmarkDurationMetric()
+                : base("Duration", "Duration", PerformanceMetricUnits.Milliseconds)
+            {
+            }
+
+            public override PerformanceMetricEvaluator CreateEvaluator(PerformanceMetricEvaluationContext context)
+            {
+                return new BenchmarkDurationEvaluator(context);
+            }
+        }
+    }
+}
